moved and redirected paths to src
