+++ src/components/Header.js
@@ -10,7 +10,7 @@ import Box from "@mui/material/Box";
 import Toolbar from "@mui/material/Toolbar";
 import Typography from "@mui/material/Typography";
 
-import logo from "../public/next.svg"
+import logo from "../../public/next.svg"
 
 // import "./Header.css";
